requirements.txt
fastapi
uvicorn
yfinance
pydantic
sqlalchemy
pymysql
apscheduler
pandas
python-dotenv
